fix paths and missing dependencies
